messages: Allow bots access messages they sent to unsubscribed private streams.
Bots can send messages to private streams they are not subscribed to. But
they were not able to access the sent messages since usermessage rows
were not generated when a bot sends message to an unsubscribed stream.
The has_message_access used to check for the existence of usermessage for
deciding whether the bot has access to access message. This issue was fixed
by allowing all the message senders access to the messages they sent.
This would would have effect only on bots since only they were allowed to
send messages to streams they don't have access to.