billing: Ask to add card instead of update when there are no saved cards.
The functionality remains the change. Only change is we are replacing
"Update card" labels with "Add card" when there are no saved cards.
This change is for realms that are not on free trial. For realms
in free trial without saved cards we already show the correct
"Add payment method" label.